% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{check_consistent}
\alias{check_consistent}
\title{Check that a data value is consistent with its prior}
\usage{
check_consistent(point, prior, p_error = 5e-04, p_warning = 0.025)
}
\arguments{
\item{point}{(numeric) Data value to check}

\item{prior}{list of distn, parama, paramb}

\item{p_error}{(numeric) Probability value outside of which we raise an error}

\item{p_warning}{(numeric) Probability value outside of which we raise a warning}
}
\value{
(c(no_error = \if{html}{\out{<boolean>}}, no_warning = \if{html}{\out{<boolean>}}))
}
\description{
Check that a data value is consistent with its prior
}
