\name{pecan.ma.summary}
\alias{pecan.ma.summary}
\title{Generate summary statitics and diagnostics for PEcAn meta.analysis}
\usage{
  pecan.ma.summary(mcmc.object, pft, outdir,
    threshold = 1.1)
}
\description{
  Generate summary statitics and diagnostics for PEcAn
  meta.analysis
}
\author{
  unknown
}

