% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_jags_columns.R
\name{rename_jags_columns}
\alias{rename_jags_columns}
\title{renames the variables within output data frame trait.data}
\usage{
rename_jags_columns(data)
}
\arguments{
\item{data}{data frame to with variables to rename}
}
\description{
renames the variables within output data frame trait.data
}
\seealso{
used with \code{\link[PEcAn.MA]{jagify}};
}
\author{
David LeBauer
}
