% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{run_meta_analysis_pft}
\alias{run_meta_analysis_pft}
\title{Complete meta-analysis workflow for a single plant functional type (PFT)}
\usage{
run_meta_analysis_pft(
  trait_data,
  priors,
  iterations,
  pft_name,
  outdir,
  random = TRUE,
  threshold = 1.2,
  use_ghs = TRUE,
  gamma_tau = 0.01
)
}
\arguments{
\item{trait_data}{(list) Named list of trait data.}

\item{priors}{(list) Named list of priors}

\item{pft_name}{(character) Name of PFT (passed to \code{pecan.ma.summary})}

\item{outdir}{(character) Path to directory where outputs will be stored.}

\item{random}{use random effects, \code{FALSE} by default}

\item{threshold}{Gelman-Rubin convergence diagnostic (MGPRF)
default = 1.2 following Bolker 2008 Ecological Models and Data in R}

\item{gamma_tau}{(numeric; default = 0.01) Prior on gamma tau parameter}
}
\value{
(list) List of trait meta-analysis results, including:
\itemize{
\item \code{trait.mcmc}: MCMC samples
\item \code{post.distns}: Posterior distributions
\item \code{jagged.data}: "JAGS-ified" input data (after GHG screen, if applied)
}
}
\description{
Complete meta-analysis workflow for a single plant functional type (PFT)
}
