% Generated by roxygen2 (4.0.2): do not edit by hand
\name{transform.nas}
\alias{transform.nas}
\title{Function to remove NA values from database queries}
\usage{
\method{transform}{nas}(data)
}
\arguments{
\item{data}{input data}
}
\value{
A data frame NAs sensibly replaced
}
\description{
Transform NA values in data exported from BETYdb
}

