% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagify.R
\name{transform_nas}
\alias{transform_nas}
\title{Function to remove NA values from database queries}
\usage{
transform_nas(data)
}
\arguments{
\item{data}{input data}
}
\value{
A data frame NAs sensibly replaced
}
\description{
Transform NA values in data exported from BETYdb
}
