\name{jagify}
\alias{jagify}
\title{jagify}
\usage{
  jagify(result)
}
\arguments{
  \item{result}{}
}
\value{
  result transformed to meet requirements of PEcAn
  meta-analysis model
}
\description{
  Convert queried data to format required by JAGS
  meta-analysis model
}
\author{
  David LeBauer
}

