% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagify.R
\name{jagify}
\alias{jagify}
\title{Prepare trait data for JAGS meta-analysis}
\usage{
jagify(result, use_ghs = TRUE)
}
\arguments{
\item{result}{input trait data}

\item{use_ghs}{(Logical) If \code{FALSE}, exclude all greenhouse data. If \code{TRUE}, use all data, including greenhouse data.}
}
\value{
result transformed to meet requirements of PEcAn meta-analysis model
}
\description{
Convert queried data to format required by JAGS meta-analysis model
}
\author{
David LeBauer
}
