% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagify.R
\name{jagify}
\alias{jagify}
\title{Prepare trait data for JAGS meta-analysis}
\usage{
jagify(result)
}
\arguments{
\item{result}{input trait data}
}
\value{
result transformed to meet requirements of PEcAn meta-analysis model
}
\description{
Convert queried data to format required by JAGS meta-analysis model
}
\author{
David LeBauer
}

