% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{meta_analysis_standalone}
\alias{meta_analysis_standalone}
\title{Complete meta-analysis workflow for a single plant functional type (PFT)}
\usage{
meta_analysis_standalone(
  trait_data,
  priors,
  iterations,
  outdir = file.path(tempdir(), "pecan-meta-analysis"),
  pft_name = NA_character_,
  random = TRUE,
  threshold = 1.2,
  use_ghs = TRUE,
  gamma_tau = 0.01
)
}
\arguments{
\item{trait_data}{(list) Named list of trait data. List item names must be
trait names (consistent with \code{priors} argument). List values are
\code{data.frame}s with the following required columns:
\code{name}, \code{mean} \code{statname}, \code{stat}, \code{greenhouse}, \code{n},
\code{site_id}, \code{specie_id}, \code{citation_id}, \code{cultivar_id},
\code{date}, \code{time}, \code{control}}

\item{priors}{(list) Named list of priors}

\item{outdir}{(character; default = \code{tempdir() / "pecan-meta-analysis"})
Path to directory where outputs will be stored.}

\item{pft_name}{(character; default = NA) Name of PFT (for logging purposes).}

\item{random}{use random effects, \code{FALSE} by default}

\item{threshold}{Gelman-Rubin convergence diagnostic (MGPRF)
default = 1.2 following Bolker 2008 Ecological Models and Data in R}

\item{gamma_tau}{(numeric; default = 0.01) Prior on gamma tau parameter}
}
\value{
(list) List of trait meta-analysis results, including:
\itemize{
\item \code{trait.mcmc}: MCMC samples
\item \code{post.distns}: Posterior distributions
\item \code{jagged.data}: "JAGS-ified" input data (after GHG screen, if applied)
}
}
\description{
Complete meta-analysis workflow for a single plant functional type (PFT)
}
