% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.analysis.R
\name{pecan.ma}
\alias{pecan.ma}
\title{Trait Meta-analysis}
\usage{
pecan.ma(
  trait.data,
  prior.distns,
  taupriors,
  j.iter,
  outdir,
  random = FALSE,
  overdispersed = TRUE,
  logfile = file.path(outdir, "meta-analysis.log)"),
  verbose = TRUE
)
}
\arguments{
\item{trait.data}{list of \code{data.frame}s, one per trait for which
data is available, generated by call to
\code{\link[PEcAn.DB:query.traits]{PEcAn.DB::query.traits()}}, and post-processed by \code{\link[=jagify]{jagify()}}.}

\item{prior.distns}{\code{data.frame} of prior distributions generated
by call to \code{\link[PEcAn.DB:query.priors]{PEcAn.DB::query.priors()}}}

\item{taupriors}{priors on variance parameters, can be scaled as needed with data mean}

\item{j.iter}{number of MCMC samples}

\item{outdir}{output directory}

\item{random}{use random effects, \code{FALSE} by default}

\item{overdispersed}{\code{TRUE} by default, if set to \code{FALSE}, data mean will be used as starting point for MCMC chains (use with caution)}

\item{logfile}{Path to file for sinking meta analysis output. If
\code{NULL}, only print output to console.}

\item{verbose}{Logical. If \code{TRUE} (default), print progress messages.}

\item{data}{data frame generated by jagify function
with indexed values for greenhouse, treatment, and site (ghs, trt, site)
as well as Y, SE, and n for each observation or summary statistic.}
}
\value{
four chains with 5000 total samples from posterior
}
\description{
Runs hierarchical meta-analysis of plant trait data
}
\details{
\code{pecan.ma} runs a hierarchical Bayesian meta-analytical model.
This model combines prior information with data from studies on the particular species or group of interest.
Data that is incorporated into the meta-analysis include the mean (\code{Y}), sample size (\code{n}),
and precision (\code{obs.prec}).
Where a set of data includes more than one level of treatment, comes from more than one site,
or comes from both field and greenhouse studies, these variables are included as random (\code{treatment}, \code{site})
or fixed (\code{greenhouse}) effects.
The \code{pecan.ma} function writes a model for each specific data set and prior using the \code{\link[=write.ma.model]{write.ma.model()}}
function to modify the \code{ma.model.template.bug} generic model.
}
\examples{
\dontrun{
  # Setup
  con <- PEcAn.DB::db.open(...)
  pft <- "temperate.Early_Hardwood"
  pft_id <- PEcAn.DB::db.query("SELECT id FROM pfts WHERE name = $1", con,
  values = list(pft))[[1]]
  traits <- c("SLA", "Vcmax")
  trait_string <- paste(shQuote(traits), collapse = ",")
  
  # Load traits and priors from BETY
  species <- PEcAn.DB::query.pft_species(pft, con = con)
  trait.data <- PEcAn.DB::query.traits(species[["id"]], c("SLA", "Vcmax"), con = con)
  prior.distns <- PEcAn.DB::query.priors(pft_id, trait_string, con = con)
  
  # Pre-process data
  jagged.data <- lapply(trait.data, PEcAn.MA::jagify)
  taupriors <- list(tauA = 0.01,
  tauB = c(SLA = 1000, Vcmax = 1000))
  result <- pecan.ma(jagged.data, prior.distns, taupriors,
                     j.iter = 5000, outdir = tempdir())
}
}
\author{
David LeBauer, Michael C. Dietze, Alexey Shiklomanov
}
