% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{run.meta.analysis.pft}
\alias{run.meta.analysis.pft}
\title{"Workflow" version of run.meta.analysis.pft}
\usage{
run.meta.analysis.pft(
  pft,
  iterations,
  random = TRUE,
  threshold = 1.2,
  dbfiles,
  dbcon,
  use_ghs = TRUE,
  update = FALSE
)
}
\arguments{
\item{random}{use random effects, \code{FALSE} by default}

\item{threshold}{Gelman-Rubin convergence diagnostic (MGPRF)
default = 1.2 following Bolker 2008 Ecological Models and Data in R}
}
\description{
Thin wrapper around \code{meta_analysis_standalone} that also reads/writes files
and registers results in the PEcAn database.
}
