\name{run.meta.analysis}
\alias{run.meta.analysis}
\title{Invoke PEcAn meta.analysis}
\usage{
  run.meta.analysis()
}
\value{
  nothing, as side effect saves \code{trait.mcmc} created
  by \code{\link{pecan.ma}} and post.distns created by
  \code{\link{approx.posterior(trait.mcmc, ...)}} to
  trait.mcmc.Rdata \ and post.distns.Rdata, respectively
}
\description{
  Run meta analysis
}
\author{
  Shawn Serbin, David LeBauer
}

