% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{run.meta.analysis}
\alias{run.meta.analysis}
\title{Run meta analysis}
\usage{
run.meta.analysis(
  pfts,
  iterations,
  random = TRUE,
  threshold = 1.2,
  dbfiles,
  database,
  use_ghs = TRUE,
  update = FALSE
)
}
\arguments{
\item{pfts}{the list of pfts to get traits for}

\item{iterations}{the number of iterations for the mcmc analysis}

\item{random}{should random effects be used?}

\item{threshold}{Gelman-Rubin convergence diagnostic, passed on to
\code{\link{pecan.ma.summary}}}

\item{dbfiles}{location where previous results are found}

\item{database}{database connection parameters}

\item{use_ghs}{do not exclude greenhouse data if TRUE}

\item{update}{logical: Rerun the meta-analysis if result files already exist?}
}
\value{
nothing, as side effect saves \code{trait.mcmc} created by
\code{\link{pecan.ma}} and post.distns created by
\code{\link{approx.posterior}(trait.mcmc, ...)}  to trait.mcmc.Rdata
and post.distns.Rdata, respectively
}
\description{
This will use the following items from settings:
\itemize{
\item settings$pfts
\item settings$database$bety
\item settings$database$dbfiles
\item settings$meta.analysis$update
}
}
\author{
Shawn Serbin, David LeBauer
}
