% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.meta.analysis.R
\name{runModule.run.meta.analysis}
\alias{runModule.run.meta.analysis}
\title{Run meta-analysis on all PFTs in a (list of) PEcAn settings}
\usage{
runModule.run.meta.analysis(settings)
}
\arguments{
\item{settings}{a PEcAn settings or MultiSettings object}
}
\value{
list of PFTs, invisibly;
saves MA results to \code{settings$pft$outdir} as a side effect
}
\description{
Run meta-analysis on all PFTs in a (list of) PEcAn settings
}
